Create balanced folds for cross validation using stratified sampling

Create balanced folds for cross validation using stratified sampling

balancedCvFolds(..., nfolds = 10)

# deprecated, please use balancedCvFolds()
balanced.cv.folds(..., nfolds = 10)

Arguments

...

vectors representing data

nfolds

number of folds to be created

Value

list with given input, nfolds and result. The result is a list matching the input with foldid attributed to each position.

Examples

balancedCvFolds(seq(10), seq(11, 15), nfolds = 2)
#> $input
#> $input[[1]]
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> $input[[2]]
#> [1] 11 12 13 14 15
#> 
#> 
#> $output
#> $output[[1]]
#>  [1] 1 1 2 2 2 1 1 2 1 2
#> 
#> $output[[2]]
#> [1] 1 2 2 1 1
#> 
#> 
#> $nfolds
#> [1] 2
#> 

# will give a warning
balancedCvFolds(seq(10), seq(11, 13), nfolds = 10)
#> Warning: Number of elements in vector (13) is less than 'nfolds' (10)
#> $input
#> $input[[1]]
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> $input[[2]]
#> [1] 11 12 13
#> 
#> 
#> $output
#> $output[[1]]
#>  [1]  5 10  4  9  2  8  7  6  1  3
#> 
#> $output[[2]]
#> [1] 2 1 3
#> 
#> 
#> $nfolds
#> [1] 10
#> 

balancedCvFolds(seq(100), seq(101, 133), nfolds = 10)
#> $input
#> $input[[1]]
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100
#> 
#> $input[[2]]
#>  [1] 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119
#> [20] 120 121 122 123 124 125 126 127 128 129 130 131 132 133
#> 
#> 
#> $output
#> $output[[1]]
#>   [1]  2  3 10  9  2 10 10  4  7  2  6  6  1  8  3  1  6  5  5  7  8  3  8  8  5
#>  [26]  5  2  3  8  8  3  6  9  9  4  4 10  8  8  9  7  2  4  1  2  6  2  9  4  6
#>  [51]  4  5  9  9  1  2  3  8  1 10  1  8  4  9  1  7 10  3  4  5 10  7  7  9  9
#>  [76]  7  2  1  3  3  7  6  4  1  7  5 10  1  7  6  3  6  4 10  5  2 10  5  5  6
#> 
#> $output[[2]]
#>  [1]  2  7  1 10  8  2  7  3 10  5  8  6  1 10  3  4  3  6  4  2  9  5  9  4  8
#> [26]  5  3  2  1  6  1  7  9
#> 
#> 
#> $nfolds
#> [1] 10
#>