Auxiliary function to generate suitable lambda parameters

Auxiliary function to generate suitable lambda parameters

buildLambda(
  lambdaLargest = NULL,
  xdata = NULL,
  ydata = NULL,
  family = NULL,
  ordersOfMagnitudeSmaller = 3,
  lambdaPerOrderMagnitude = 150,
  lambda.largest = deprecated(),
  orders.of.magnitude.smaller = deprecated(),
  lambda.per.order.magnitude = deprecated()
)

Arguments

lambdaLargest

numeric value for largest number of lambda to consider (usually with a target of 1 selected variable)

xdata

X parameter for glmnet function

ydata

Y parameter for glmnet function

family

family parameter to glmnet function

ordersOfMagnitudeSmaller

minimum value for lambda (lambda.largest / 10^orders.of.magnitude.smaller)

lambdaPerOrderMagnitude

how many lambdas to create for each order of magnitude

lambda.largest

orders.of.magnitude.smaller

lambda.per.order.magnitude

Value

a numeric vector with suitable lambdas

Examples

buildLambda(5.4)
#>   [1] 5.40000 5.36400 5.32800 5.29200 5.25600 5.22000 5.18400 5.14800 5.11200
#>  [10] 5.07600 5.04000 5.00400 4.96800 4.93200 4.89600 4.86000 4.82400 4.78800
#>  [19] 4.75200 4.71600 4.68000 4.64400 4.60800 4.57200 4.53600 4.50000 4.46400
#>  [28] 4.42800 4.39200 4.35600 4.32000 4.28400 4.24800 4.21200 4.17600 4.14000
#>  [37] 4.10400 4.06800 4.03200 3.99600 3.96000 3.92400 3.88800 3.85200 3.81600
#>  [46] 3.78000 3.74400 3.70800 3.67200 3.63600 3.60000 3.56400 3.52800 3.49200
#>  [55] 3.45600 3.42000 3.38400 3.34800 3.31200 3.27600 3.24000 3.20400 3.16800
#>  [64] 3.13200 3.09600 3.06000 3.02400 2.98800 2.95200 2.91600 2.88000 2.84400
#>  [73] 2.80800 2.77200 2.73600 2.70000 2.66400 2.62800 2.59200 2.55600 2.52000
#>  [82] 2.48400 2.44800 2.41200 2.37600 2.34000 2.30400 2.26800 2.23200 2.19600
#>  [91] 2.16000 2.12400 2.08800 2.05200 2.01600 1.98000 1.94400 1.90800 1.87200
#> [100] 1.83600 1.80000 1.76400 1.72800 1.69200 1.65600 1.62000 1.58400 1.54800
#> [109] 1.51200 1.47600 1.44000 1.40400 1.36800 1.33200 1.29600 1.26000 1.22400
#> [118] 1.18800 1.15200 1.11600 1.08000 1.04400 1.00800 0.97200 0.93600 0.90000
#> [127] 0.86400 0.82800 0.79200 0.75600 0.72000 0.68400 0.64800 0.61200 0.57600
#> [136] 0.54000 0.53640 0.53280 0.52920 0.52560 0.52200 0.51840 0.51480 0.51120
#> [145] 0.50760 0.50400 0.50400 0.50040 0.49680 0.49320 0.48960 0.48600 0.48240
#> [154] 0.47880 0.47520 0.47160 0.46800 0.46800 0.46440 0.46080 0.45720 0.45360
#> [163] 0.45000 0.44640 0.44280 0.43920 0.43560 0.43200 0.42840 0.42480 0.42120
#> [172] 0.41760 0.41400 0.41040 0.40680 0.40320 0.39960 0.39600 0.39600 0.39240
#> [181] 0.38880 0.38520 0.38160 0.37800 0.37440 0.37080 0.36720 0.36360 0.36000
#> [190] 0.36000 0.35640 0.35280 0.34920 0.34560 0.34200 0.33840 0.33480 0.33120
#> [199] 0.32760 0.32400 0.32040 0.31680 0.31320 0.30960 0.30600 0.30240 0.29880
#> [208] 0.29520 0.29160 0.28800 0.28800 0.28440 0.28080 0.27720 0.27360 0.27000
#> [217] 0.26640 0.26280 0.25920 0.25560 0.25200 0.24840 0.24480 0.24120 0.23760
#> [226] 0.23400 0.23040 0.22680 0.22320 0.21960 0.21600 0.21240 0.20880 0.20520
#> [235] 0.20160 0.19800 0.19440 0.19080 0.18720 0.18360 0.18000 0.17640 0.17280
#> [244] 0.16920 0.16560 0.16200 0.15840 0.15480 0.15120 0.14760 0.14400 0.14040
#> [253] 0.13680 0.13320 0.12960 0.12600 0.12240 0.11880 0.11520 0.11160 0.10800
#> [262] 0.10440 0.10080 0.09720 0.09360 0.09000 0.08640 0.08280 0.07920 0.07560
#> [271] 0.07200 0.06840 0.06480 0.06120 0.05760 0.05400 0.05364 0.05328 0.05292
#> [280] 0.05256 0.05220 0.05184 0.05148 0.05112 0.05076 0.05040 0.05040 0.05004
#> [289] 0.04968 0.04932 0.04896 0.04860 0.04824 0.04788 0.04752 0.04716 0.04680
#> [298] 0.04680 0.04644 0.04608 0.04572 0.04536 0.04500 0.04464 0.04428 0.04392
#> [307] 0.04356 0.04320 0.04320 0.04284 0.04248 0.04212 0.04176 0.04140 0.04104
#> [316] 0.04068 0.04032 0.03996 0.03960 0.03924 0.03888 0.03852 0.03816 0.03780
#> [325] 0.03744 0.03708 0.03672 0.03636 0.03600 0.03564 0.03528 0.03492 0.03456
#> [334] 0.03420 0.03384 0.03348 0.03312 0.03276 0.03240 0.03204 0.03168 0.03132
#> [343] 0.03096 0.03060 0.03024 0.02988 0.02952 0.02916 0.02880 0.02880 0.02844
#> [352] 0.02808 0.02772 0.02736 0.02700 0.02664 0.02628 0.02592 0.02556 0.02520
#> [361] 0.02484 0.02448 0.02412 0.02376 0.02340 0.02304 0.02268 0.02232 0.02196
#> [370] 0.02160 0.02160 0.02124 0.02088 0.02052 0.02016 0.01980 0.01944 0.01908
#> [379] 0.01872 0.01836 0.01800 0.01764 0.01728 0.01692 0.01656 0.01620 0.01584
#> [388] 0.01548 0.01512 0.01476 0.01440 0.01440 0.01404 0.01368 0.01332 0.01296
#> [397] 0.01260 0.01224 0.01188 0.01152 0.01116 0.01080 0.01080 0.01044 0.01008
#> [406] 0.00972 0.00936 0.00900 0.00864 0.00828 0.00792 0.00756 0.00720 0.00720
#> [415] 0.00684 0.00648 0.00612 0.00576 0.00540 0.00504 0.00468 0.00432 0.00396
#> [424] 0.00360 0.00360 0.00324 0.00288 0.00252 0.00216 0.00180 0.00144 0.00108
#> [433] 0.00072 0.00036