Build gene network from peptide ids

This can reduce the dimension of the original network, as there may not be a mapping between peptide and gene id

buildStringNetwork(
  stringTbl,
  useNames = c("protein", "ensembl", "external"),
  string.tbl = deprecated(),
  use.names = deprecated()
)

Arguments

stringTbl

data.frame or tibble with colnames and rownames as ensembl peptide id (same order).

useNames

character(1) that defaults to use protein names _('protein'), other options are 'ensembl' for ensembl gene id or 'external' for external gene names.

string.tbl

use.names

Value

a new matrix with gene ids instead of peptide ids. The size of matrix can be different as there may not be a mapping or a peptide mapping can have multiple genes.

See also

Examples

# \donttest{
interactions <- stringDBhomoSapiens(scoreThreshold = 100)
#> STRINGdb download: This might take some time, please be patient.
#> Rows: 11759454 Columns: 16
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: " "
#> chr  (2): protein1, protein2
#> dbl (14): neighborhood, neighborhood_transferred, fusion, cooccurence, homol...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
string_network <- buildStringNetwork(interactions)

# number of edges
sum(string_network != 0)
#> [1] 5592508
# }