Calculate cross validating GLM model with network-based regularization

network parameter accepts:

cv.glmDegree(
  xdata,
  ydata,
  network,
  options = networkOptions(),
  experiment = NULL,
  network.options = deprecated(),
  experiment.name = deprecated(),
  ...
)

cv.glmHub(
  xdata,
  ydata,
  network,
  options = networkOptions(),
  experiment = NULL,
  network.options = deprecated(),
  experiment.name = deprecated(),
  ...
)

cv.glmOrphan(
  xdata,
  ydata,
  network,
  options = networkOptions(),
  experiment = NULL,
  network.options = deprecated(),
  experiment.name = deprecated(),
  ...
)

cv.glmSparseNet(
  xdata,
  ydata,
  network,
  options = networkOptions(),
  experiment = NULL,
  network.options = deprecated(),
  experiment.name = deprecated(),
  ...
)

Arguments

xdata

input data, can be a matrix or MultiAssayExperiment.

ydata

response data compatible with glmnet.

network

type of network, see below.

options

options to calculate network.

experiment

name of experiment to use as input in MultiAssayExperiment object (only if xdata is an object of this class).

network.options

experiment.name

...

parameters that glmnet::cv.glmnet() accepts.

Value

an object just as cv.glmnet

Details

string to calculate network based on data (correlation, covariance)

matrix representing the network

vector with already calculated penalty weights (can also be used directly glmnet)

Functions

cv.glmDegree(): penalizes nodes with small degree (inversion penalization h(x) = 1 / x).

cv.glmHub(): penalizes nodes with small degree (normalized heuristic that promotes nodes with many edges).

cv.glmOrphan(): penalizes nodes with high degree (normalized heuristic that promotes nodes with few edges).

See also

Examples

# Degree penalization

xdata <- matrix(rnorm(100), ncol = 5)
cv.glmDegree(
    xdata,
    rnorm(nrow(xdata)),
    "correlation",
    family = "gaussian",
    nfolds = 5,
    options = networkOptions(minDegree = .2)
)
#> Warning: Could not create cache base folder at '/tmp/RtmpNd8cjn/R.INSTALLb9463dc086a7/glmSparseNet/run-cache'... trying to use current working directory
#> Warning: Could not create cache base folder at '/tmp/RtmpNd8cjn/R.INSTALLb9463dc086a7/glmSparseNet/run-cache'... trying to use current working directory
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, nfolds = 5, penalty.factor = penaltyFactor,      family = "gaussian") 
#> 
#> Measure: Mean-Squared Error 
#> 
#>     Lambda Index Measure     SE Nonzero
#> min 0.3094     5   1.357 0.6439       3
#> 1se 0.4489     1   1.375 0.6425       0
# Hub penalization

xdata <- matrix(rnorm(100), ncol = 5)
cv.glmHub(
    xdata,
    rnorm(nrow(xdata)),
    "correlation",
    family = "gaussian",
    nfolds = 5,
    options = networkOptions(minDegree = .2)
)
#> Warning: Could not create cache base folder at '/tmp/RtmpNd8cjn/R.INSTALLb9463dc086a7/glmSparseNet/run-cache'... trying to use current working directory
#> Warning: Could not create cache base folder at '/tmp/RtmpNd8cjn/R.INSTALLb9463dc086a7/glmSparseNet/run-cache'... trying to use current working directory
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, nfolds = 5, penalty.factor = penaltyFactor,      family = "gaussian") 
#> 
#> Measure: Mean-Squared Error 
#> 
#>     Lambda Index Measure     SE Nonzero
#> min 0.3303     1  0.5652 0.1803       0
#> 1se 0.3303     1  0.5652 0.1803       0
# Orphan penalization

xdata <- matrix(rnorm(100), ncol = 5)
cv.glmOrphan(
    xdata,
    rnorm(nrow(xdata)),
    "correlation",
    family = "gaussian",
    nfolds = 5,
    options = networkOptions(minDegree = .2)
)
#> Warning: Could not create cache base folder at '/tmp/RtmpNd8cjn/R.INSTALLb9463dc086a7/glmSparseNet/run-cache'... trying to use current working directory
#> Warning: Could not create cache base folder at '/tmp/RtmpNd8cjn/R.INSTALLb9463dc086a7/glmSparseNet/run-cache'... trying to use current working directory
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, nfolds = 5, penalty.factor = penaltyFactor,      family = "gaussian") 
#> 
#> Measure: Mean-Squared Error 
#> 
#>     Lambda Index Measure    SE Nonzero
#> min 0.2323     1    1.39 0.408       0
#> 1se 0.2323     1    1.39 0.408       0
# \donttest{
# Gaussian model
xdata <- matrix(rnorm(500), ncol = 5)
cv.glmSparseNet(
    xdata, rnorm(nrow(xdata)), "correlation",
    family = "gaussian"
)
#> Warning: Could not create cache base folder at '/tmp/RtmpNd8cjn/R.INSTALLb9463dc086a7/glmSparseNet/run-cache'... trying to use current working directory
#> Warning: Could not create cache base folder at '/tmp/RtmpNd8cjn/R.INSTALLb9463dc086a7/glmSparseNet/run-cache'... trying to use current working directory
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, penalty.factor = penaltyFactor,      family = "gaussian") 
#> 
#> Measure: Mean-Squared Error 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.08637    10   1.167 0.1770       3
#> 1se 0.19952     1   1.175 0.1772       0
cv.glmSparseNet(
    xdata, rnorm(nrow(xdata)), "covariance",
    family = "gaussian"
)
#> Warning: Could not create cache base folder at '/tmp/RtmpNd8cjn/R.INSTALLb9463dc086a7/glmSparseNet/run-cache'... trying to use current working directory
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, penalty.factor = penaltyFactor,      family = "gaussian") 
#> 
#> Measure: Mean-Squared Error 
#> 
#>     Lambda Index Measure     SE Nonzero
#> min 0.1754     1  0.9548 0.1459       0
#> 1se 0.1754     1  0.9548 0.1459       0
# }
#
#
# Using MultiAssayExperiment with survival model
library(MultiAssayExperiment)
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
data("miniACC", package = "MultiAssayExperiment")

xdata <- miniACC

#
# build valid data with days of last follow up or to event
event.ix <- which(!is.na(xdata$days_to_death))
cens.ix <- which(!is.na(xdata$days_to_last_followup))
xdata$surv_event_time <- array(NA, nrow(colData(xdata)))
xdata$surv_event_time[event.ix] <- xdata$days_to_death[event.ix]
xdata$surv_event_time[cens.ix] <- xdata$days_to_last_followup[cens.ix]

#
# Keep only valid individuals
valid.ix <- as.vector(!is.na(xdata$surv_event_time) &
    !is.na(xdata$vital_status) &
    xdata$surv_event_time > 0)
xdata.valid <- xdata[, rownames(colData(xdata))[valid.ix]]
ydata.valid <- colData(xdata.valid)[, c("surv_event_time", "vital_status")]
colnames(ydata.valid) <- c("time", "status")

#
cv.glmSparseNet(
    xdata.valid,
    ydata.valid,
    nfolds     = 5,
    family     = "cox",
    network    = "correlation",
    experiment = "RNASeq2GeneNorm"
)
#> Warning: 'experiments' dropped; see 'drops()'
#> harmonizing input:
#>   removing 304 sampleMap rows not in names(experiments)
#>   removing 12 colData rownames not in sampleMap 'primary'
#> Warning: Could not create cache base folder at '/tmp/RtmpNd8cjn/R.INSTALLb9463dc086a7/glmSparseNet/run-cache'... trying to use current working directory
#> Warning: Could not create cache base folder at '/tmp/RtmpNd8cjn/R.INSTALLb9463dc086a7/glmSparseNet/run-cache'... trying to use current working directory
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, nfolds = 5, penalty.factor = penaltyFactor,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>     Lambda Index Measure     SE Nonzero
#> min 0.1189    26   8.261 0.5460      13
#> 1se 0.2747     8   8.756 0.3704       2