Run function and save cache

This method saves the function that's being called

.runCache(
  fun,
  ...,
  seed = NULL,
  baseDir = NULL,
  cachePrefix = "generic_cache",
  cacheDigest = list(),
  showMessage = NULL,
  forceRecalc = FALSE,
  addToHash = NULL
)

# S4 method for function
.runCache(
  fun,
  ...,
  seed = NULL,
  baseDir = NULL,
  cachePrefix = "generic_cache",
  cacheDigest = list(),
  showMessage = NULL,
  forceRecalc = FALSE,
  addToHash = NULL
)

Arguments

fun

function call name

...

parameters for function call

seed

when function call is random, this allows to set seed beforehand

baseDir

directory where data is stored

cachePrefix

prefix for file name to be generated from parameters (...)

cacheDigest

cache of the digest for one or more of the parameters

showMessage

show message that data is being retrieved from cache

forceRecalc

force the recalculation of the values

addToHash

something to add to the filename generation

Value

the result of fun(...)

Functions

.runCache(`function`): accepts function as first argument and save cache

Examples

# [optional] save cache in a temporary directory
#
glmSparseNet:::.baseDir(tempdir())
#> [1] "/tmp/RtmpWbkmJb"
glmSparseNet:::.runCache(c, 1, 2, 3, 4)
#> Saving in cache:  /tmp/RtmpWbkmJb/9a54/cache-generic_cache-H_9a54ea6e9ee52dc8bdcbc4f79272ee77.RData
#> [1] 1 2 3 4
#
# next three should use the same cache
#  note, the middle call should be a little faster as digest is not
#  calculated
#   for the first argument
glmSparseNet:::.runCache(c, 1, 2, 3, 4)
#> Loading from cache (not calculating):
#>   /tmp/RtmpWbkmJb/9a54/cache-generic_cache-H_9a54ea6e9ee52dc8bdcbc4f79272ee77.RData
#> Cache was created at 2024-02-14 12:30:51.38195 using glmSparseNet v1.21.2
#> [1] 1 2 3 4
glmSparseNet:::.runCache(c, a = 1, 2, c = 3, 4)
#> Loading from cache (not calculating):
#>   /tmp/RtmpWbkmJb/9a54/cache-generic_cache-H_9a54ea6e9ee52dc8bdcbc4f79272ee77.RData
#> Cache was created at 2024-02-14 12:30:51.38195 using glmSparseNet v1.21.2
#> [1] 1 2 3 4
# \donttest{
# Using a local folder
# glmSparseNet:::.runCache(c, 1, 2, 3, 4, baseDir = "runcache")
# }