Calculate GLM model with network-based regularization

network parameter accepts:

string to calculate network based on data (correlation, covariance)

matrix representing the network

vector with already calculated penalty weights (can also be used directly with glmnet)

glmSparseNet(
  xdata,
  ydata,
  network,
  options = networkOptions(),
  experiment = NULL,
  network.options = deprecated(),
  experiment.name = deprecated(),
  ...
)

glmDegree(
  xdata,
  ydata,
  network,
  options = networkOptions(),
  experiment = NULL,
  network.options = deprecated(),
  experiment.name = deprecated(),
  ...
)

glmHub(
  xdata,
  ydata,
  network,
  options = networkOptions(),
  experiment = NULL,
  network.options = deprecated(),
  experiment.name = deprecated(),
  ...
)

glmOrphan(
  xdata,
  ydata,
  network,
  options = networkOptions(),
  experiment = NULL,
  network.options = deprecated(),
  experiment.name = deprecated(),
  ...
)

Arguments

xdata

input data, can be a matrix or MultiAssayExperiment.

ydata

response data compatible with glmnet.

network

type of network, see below.

options

options to calculate network.

experiment

name of experiment to use as input in MultiAssayExperiment object (only if xdata is an object of this class).

network.options

experiment.name

...

parameters that glmnet::glmnet() accepts.

Value

an object just as glmnet

Functions

glmDegree(): penalizes nodes with small degree (inversion penalization h(x) = 1 / x).

glmHub(): Penalizes nodes with small degree (normalized heuristic that promotes nodes with many edges).

glmOrphan(): Penalizes nodes with high degree (normalized heuristic that promotes nodes with few edges).

See also

Examples

xdata <- matrix(rnorm(100), ncol = 20)
glmSparseNet(xdata, rnorm(nrow(xdata)), "correlation", family = "gaussian")
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, family = "gaussian", penalty.factor = penaltyFactor) 
#> 
#>    Df  %Dev  Lambda
#> 1   0  0.00 0.56300
#> 2   1  7.62 0.53740
#> 3   1 14.56 0.51290
#> 4   1 20.89 0.48960
#> 5   1 26.65 0.46740
#> 6   1 31.90 0.44610
#> 7   1 36.69 0.42590
#> 8   1 41.05 0.40650
#> 9   2 45.23 0.38800
#> 10  2 49.33 0.37040
#> 11  2 53.07 0.35360
#> 12  2 56.48 0.33750
#> 13  2 59.58 0.32210
#> 14  2 62.41 0.30750
#> 15  2 64.99 0.29350
#> 16  2 67.34 0.28020
#> 17  2 69.48 0.26740
#> 18  3 71.63 0.25530
#> 19  3 74.14 0.24370
#> 20  3 76.42 0.23260
#> 21  3 78.49 0.22200
#> 22  3 80.39 0.21190
#> 23  3 82.11 0.20230
#> 24  3 83.68 0.19310
#> 25  3 85.11 0.18430
#> 26  3 86.42 0.17600
#> 27  3 87.61 0.16800
#> 28  3 88.69 0.16030
#> 29  3 89.68 0.15300
#> 30  4 90.58 0.14610
#> 31  4 91.41 0.13940
#> 32  4 92.18 0.13310
#> 33  4 92.87 0.12710
#> 34  4 93.50 0.12130
#> 35  4 94.08 0.11580
#> 36  4 94.60 0.11050
#> 37  4 95.08 0.10550
#> 38  4 95.52 0.10070
#> 39  4 95.92 0.09612
#> 40  4 96.28 0.09175
#> 41  4 96.61 0.08758
#> 42  4 96.91 0.08360
#> 43  4 97.19 0.07980
#> 44  4 97.44 0.07617
#> 45  4 97.66 0.07271
#> 46  4 97.87 0.06940
#> 47  4 98.06 0.06625
#> 48  4 98.23 0.06324
#> 49  4 98.39 0.06036
#> 50  4 98.53 0.05762
#> 51  4 98.66 0.05500
#> 52  4 98.78 0.05250
#> 53  4 98.89 0.05012
#> 54  4 98.99 0.04784
#> 55  4 99.08 0.04566
#> 56  4 99.16 0.04359
#> 57  4 99.23 0.04161
#> 58  4 99.30 0.03972
#> 59  4 99.36 0.03791
#> 60  4 99.42 0.03619
#> 61  4 99.47 0.03454
#> 62  4 99.52 0.03297
#> 63  4 99.56 0.03147
#> 64  4 99.60 0.03004
#> 65  4 99.64 0.02868
#> 66  4 99.67 0.02737
#> 67  4 99.70 0.02613
#> 68  4 99.72 0.02494
#> 69  4 99.75 0.02381
#> 70  4 99.77 0.02273
#> 71  4 99.79 0.02169
#> 72  4 99.81 0.02071
#> 73  4 99.83 0.01977
#> 74  4 99.84 0.01887
#> 75  4 99.86 0.01801
#> 76  4 99.87 0.01719
#> 77  4 99.88 0.01641
#> 78  4 99.89 0.01566
#> 79  4 99.90 0.01495
glmSparseNet(xdata, rnorm(nrow(xdata)), "covariance", family = "gaussian")
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, family = "gaussian", penalty.factor = penaltyFactor) 
#> 
#>    Df  %Dev  Lambda
#> 1   0  0.00 0.98040
#> 2   1  6.60 0.93590
#> 3   1 12.62 0.89330
#> 4   2 18.29 0.85270
#> 5   2 25.52 0.81400
#> 6   2 32.10 0.77700
#> 7   2 38.09 0.74170
#> 8   2 43.56 0.70790
#> 9   2 48.53 0.67580
#> 10  2 53.07 0.64500
#> 11  2 57.20 0.61570
#> 12  2 60.97 0.58770
#> 13  2 64.40 0.56100
#> 14  2 67.52 0.53550
#> 15  2 70.37 0.51120
#> 16  2 72.97 0.48800
#> 17  2 75.33 0.46580
#> 18  2 77.49 0.44460
#> 19  2 79.45 0.42440
#> 20  2 81.24 0.40510
#> 21  2 82.87 0.38670
#> 22  2 84.35 0.36910
#> 23  2 85.71 0.35230
#> 24  2 86.94 0.33630
#> 25  2 88.06 0.32100
#> 26  2 89.09 0.30650
#> 27  2 90.02 0.29250
#> 28  2 90.87 0.27920
#> 29  2 91.64 0.26650
#> 30  3 92.35 0.25440
#> 31  3 93.00 0.24290
#> 32  3 93.60 0.23180
#> 33  3 94.14 0.22130
#> 34  3 94.64 0.21120
#> 35  3 95.09 0.20160
#> 36  3 95.49 0.19250
#> 37  3 95.87 0.18370
#> 38  4 96.23 0.17540
#> 39  4 96.56 0.16740
#> 40  4 96.87 0.15980
#> 41  4 97.15 0.15250
#> 42  4 97.40 0.14560
#> 43  4 97.63 0.13900
#> 44  4 97.84 0.13270
#> 45  4 98.03 0.12660
#> 46  4 98.21 0.12090
#> 47  4 98.37 0.11540
#> 48  4 98.51 0.11010
#> 49  4 98.64 0.10510
#> 50  4 98.76 0.10030
#> 51  4 98.87 0.09579
#> 52  4 98.97 0.09143
#> 53  4 99.07 0.08728
#> 54  4 99.15 0.08331
#> 55  4 99.22 0.07952
#> 56  4 99.29 0.07591
#> 57  4 99.36 0.07246
#> 58  4 99.41 0.06917
#> 59  4 99.47 0.06602
#> 60  4 99.51 0.06302
#> 61  4 99.56 0.06016
#> 62  4 99.60 0.05742
#> 63  4 99.63 0.05481
#> 64  4 99.66 0.05232
#> 65  4 99.69 0.04994
#> 66  4 99.72 0.04767
#> 67  4 99.75 0.04551
#> 68  4 99.77 0.04344
#> 69  4 99.79 0.04146
#> 70  4 99.81 0.03958
#> 71  4 99.82 0.03778
#> 72  4 99.84 0.03606
#> 73  4 99.85 0.03442
#> 74  4 99.87 0.03286
#> 75  4 99.88 0.03137
#> 76  4 99.89 0.02994
#> 77  4 99.90 0.02858
#> 78  4 99.91 0.02728

#
#
# Using MultiAssayExperiment
# load data
library(MultiAssayExperiment)
data("miniACC", package = "MultiAssayExperiment")

xdata <- miniACC
# TODO aking out x individuals missing values
# build valid data with days of last follow up or to event
event.ix <- which(!is.na(xdata$days_to_death))
cens.ix <- which(!is.na(xdata$days_to_last_followup))

xdata$surv_event_time <- array(NA, nrow(colData(xdata)))
xdata$surv_event_time[event.ix] <- xdata$days_to_death[event.ix]
xdata$surv_event_time[cens.ix] <- xdata$days_to_last_followup[cens.ix]

# Keep only valid individuals
valid.ix <- as.vector(!is.na(xdata$surv_event_time) &
    !is.na(xdata$vital_status) &
    xdata$surv_event_time > 0)
xdata.valid <- xdata[, rownames(colData(xdata))[valid.ix]]
ydata.valid <- colData(xdata.valid)[, c("surv_event_time", "vital_status")]
colnames(ydata.valid) <- c("time", "status")

glmSparseNet(
    xdata.valid,
    ydata.valid,
    family = "cox",
    network = "correlation",
    experiment = "RNASeq2GeneNorm"
)
#> Warning: 'experiments' dropped; see 'drops()'
#> harmonizing input:
#>   removing 304 sampleMap rows not in names(experiments)
#>   removing 12 colData rownames not in sampleMap 'primary'
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, family = "cox", penalty.factor = penaltyFactor) 
#> 
#>     Df  %Dev  Lambda
#> 1    0  0.00 0.38050
#> 2    1  2.36 0.36320
#> 3    1  4.32 0.34670
#> 4    1  5.97 0.33090
#> 5    1  7.38 0.31590
#> 6    2  8.68 0.30150
#> 7    2  9.89 0.28780
#> 8    2 10.93 0.27480
#> 9    2 11.84 0.26230
#> 10   3 12.69 0.25030
#> 11   3 13.51 0.23900
#> 12   3 14.23 0.22810
#> 13   3 14.87 0.21770
#> 14   4 15.58 0.20780
#> 15   4 16.31 0.19840
#> 16   5 17.00 0.18940
#> 17   6 17.88 0.18080
#> 18   6 18.79 0.17260
#> 19   6 19.61 0.16470
#> 20   7 20.52 0.15720
#> 21   8 21.38 0.15010
#> 22   9 22.24 0.14330
#> 23  10 23.15 0.13670
#> 24  11 24.08 0.13050
#> 25  11 24.99 0.12460
#> 26  13 26.09 0.11890
#> 27  13 27.12 0.11350
#> 28  14 28.16 0.10840
#> 29  14 29.11 0.10340
#> 30  14 29.99 0.09874
#> 31  16 30.96 0.09425
#> 32  16 32.03 0.08997
#> 33  19 33.17 0.08588
#> 34  23 34.48 0.08198
#> 35  23 35.76 0.07825
#> 36  24 37.04 0.07469
#> 37  24 38.35 0.07130
#> 38  24 39.58 0.06806
#> 39  25 40.95 0.06497
#> 40  25 42.32 0.06201
#> 41  25 43.73 0.05919
#> 42  26 45.06 0.05650
#> 43  26 46.34 0.05394
#> 44  26 47.55 0.05148
#> 45  27 48.72 0.04914
#> 46  28 49.85 0.04691
#> 47  28 51.11 0.04478
#> 48  30 52.40 0.04274
#> 49  32 53.74 0.04080
#> 50  33 55.10 0.03895
#> 51  31 56.39 0.03718
#> 52  31 57.57 0.03549
#> 53  32 58.75 0.03387
#> 54  31 59.88 0.03233
#> 55  30 60.99 0.03086
#> 56  31 62.05 0.02946
#> 57  33 63.12 0.02812
#> 58  34 64.19 0.02684
#> 59  34 65.26 0.02562
#> 60  34 66.30 0.02446
#> 61  35 67.31 0.02335
#> 62  34 68.29 0.02229
#> 63  35 69.23 0.02127
#> 64  36 70.16 0.02031
#> 65  36 71.06 0.01938
#> 66  36 71.94 0.01850
#> 67  37 72.81 0.01766
#> 68  35 73.65 0.01686
#> 69  34 74.47 0.01609
#> 70  36 75.26 0.01536
#> 71  36 76.03 0.01466
#> 72  36 76.78 0.01400
#> 73  36 77.50 0.01336
#> 74  37 78.22 0.01275
#> 75  38 78.92 0.01217
#> 76  38 79.61 0.01162
#> 77  37 80.28 0.01109
#> 78  37 80.93 0.01059
#> 79  37 81.56 0.01011
#> 80  37 82.18 0.00965
#> 81  37 82.78 0.00921
#> 82  38 83.36 0.00879
#> 83  39 83.93 0.00839
#> 84  39 84.50 0.00801
#> 85  39 85.04 0.00764
#> 86  39 85.56 0.00730
#> 87  38 86.07 0.00697
#> 88  37 86.55 0.00665
#> 89  38 87.03 0.00635
#> 90  38 87.49 0.00606
#> 91  38 87.94 0.00578
#> 92  39 88.38 0.00552
#> 93  39 88.81 0.00527
#> 94  39 89.22 0.00503
#> 95  38 89.61 0.00480
#> 96  38 90.00 0.00458
#> 97  38 90.38 0.00438
#> 98  38 90.75 0.00418
#> 99  38 91.09 0.00399
#> 100 37 91.43 0.00380
# Degree penalization

xdata <- matrix(rnorm(100), ncol = 5)
glmDegree(
    xdata,
    rnorm(nrow(xdata)),
    "correlation",
    family = "gaussian",
    options = networkOptions(minDegree = .2)
)
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, family = "gaussian", penalty.factor = penaltyFactor) 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.286300
#> 2   1  2.00 0.260900
#> 3   1  3.65 0.237700
#> 4   1  5.03 0.216600
#> 5   2  6.59 0.197400
#> 6   2  7.95 0.179800
#> 7   2  9.08 0.163800
#> 8   2 10.01 0.149300
#> 9   3 11.46 0.136000
#> 10  3 12.90 0.123900
#> 11  3 14.10 0.112900
#> 12  3 15.10 0.102900
#> 13  3 15.93 0.093760
#> 14  3 16.61 0.085430
#> 15  3 17.18 0.077840
#> 16  3 17.66 0.070930
#> 17  3 18.05 0.064630
#> 18  3 18.38 0.058880
#> 19  3 18.65 0.053650
#> 20  4 19.00 0.048890
#> 21  4 19.30 0.044540
#> 22  4 19.55 0.040590
#> 23  4 19.76 0.036980
#> 24  4 19.94 0.033700
#> 25  4 20.08 0.030700
#> 26  5 20.21 0.027970
#> 27  5 20.34 0.025490
#> 28  5 20.44 0.023230
#> 29  5 20.53 0.021160
#> 30  5 20.60 0.019280
#> 31  5 20.66 0.017570
#> 32  5 20.71 0.016010
#> 33  5 20.75 0.014590
#> 34  5 20.78 0.013290
#> 35  5 20.81 0.012110
#> 36  5 20.83 0.011030
#> 37  5 20.85 0.010050
#> 38  5 20.87 0.009160
#> 39  5 20.88 0.008347
#> 40  5 20.89 0.007605
#> 41  5 20.90 0.006930
#> 42  5 20.91 0.006314
#> 43  5 20.92 0.005753
#> 44  5 20.92 0.005242
#> 45  5 20.93 0.004776
#> 46  5 20.93 0.004352
#> 47  5 20.93 0.003965
#> 48  5 20.93 0.003613
#> 49  5 20.94 0.003292
#> 50  5 20.94 0.003000
#> 51  5 20.94 0.002733
#> 52  5 20.94 0.002490
#> 53  5 20.94 0.002269
#> 54  5 20.94 0.002068
#> 55  5 20.94 0.001884
#> 56  5 20.94 0.001716
#> 57  5 20.94 0.001564
#> 58  5 20.94 0.001425
#> 59  5 20.95 0.001298
#> 60  5 20.95 0.001183
#> 61  5 20.95 0.001078
#> 62  5 20.95 0.000982
xdata <- matrix(rnorm(100), ncol = 5)
glmHub(
    xdata,
    rnorm(nrow(xdata)),
    "correlation",
    family = "gaussian",
    options = networkOptions(minDegree = .2)
)
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, family = "gaussian", penalty.factor = penaltyFactor) 
#> 
#>    Df  %Dev  Lambda
#> 1   0  0.00 0.46270
#> 2   1  4.22 0.42160
#> 3   1  7.72 0.38410
#> 4   1 10.63 0.35000
#> 5   1 13.05 0.31890
#> 6   1 15.05 0.29060
#> 7   1 16.71 0.26480
#> 8   1 18.10 0.24120
#> 9   1 19.24 0.21980
#> 10  1 20.19 0.20030
#> 11  1 20.99 0.18250
#> 12  1 21.64 0.16630
#> 13  1 22.19 0.15150
#> 14  1 22.64 0.13800
#> 15  2 23.05 0.12580
#> 16  2 23.79 0.11460
#> 17  3 24.54 0.10440
#> 18  3 25.23 0.09515
#> 19  3 25.80 0.08669
#> 20  3 26.28 0.07899
#> 21  3 26.68 0.07198
#> 22  3 27.01 0.06558
#> 23  3 27.28 0.05976
#> 24  4 27.57 0.05445
#> 25  4 27.92 0.04961
#> 26  4 28.21 0.04520
#> 27  4 28.45 0.04119
#> 28  4 28.65 0.03753
#> 29  4 28.82 0.03419
#> 30  4 28.95 0.03116
#> 31  5 29.09 0.02839
#> 32  5 29.20 0.02587
#> 33  5 29.30 0.02357
#> 34  5 29.38 0.02147
#> 35  5 29.45 0.01957
#> 36  5 29.50 0.01783
#> 37  5 29.55 0.01624
#> 38  5 29.59 0.01480
#> 39  5 29.62 0.01349
#> 40  5 29.64 0.01229
#> 41  5 29.67 0.01120
#> 42  5 29.68 0.01020
#> 43  5 29.70 0.00930
#> 44  5 29.71 0.00847
#> 45  5 29.72 0.00772
#> 46  5 29.73 0.00703
#> 47  5 29.74 0.00641
#> 48  5 29.74 0.00584
#> 49  5 29.75 0.00532
#> 50  5 29.75 0.00485
#> 51  5 29.76 0.00442
#> 52  5 29.76 0.00402
#> 53  5 29.76 0.00367
#> 54  5 29.76 0.00334
#> 55  5 29.77 0.00304
#> 56  5 29.77 0.00277
#> 57  5 29.77 0.00253
#> 58  5 29.77 0.00230
#> 59  5 29.77 0.00210
#> 60  5 29.77 0.00191
#> 61  5 29.77 0.00174
#> 62  5 29.77 0.00159
#> 63  5 29.77 0.00145
#> 64  5 29.77 0.00132
#> 65  5 29.77 0.00120
# Orphan penalization

xdata <- matrix(rnorm(100), ncol = 5)
glmOrphan(
    xdata,
    rnorm(nrow(xdata)),
    "correlation",
    family = "gaussian",
    options = networkOptions(minDegree = .2)
)
#> 
#> Call:  fun(x = xdataNorm, y = ydataNorm, family = "gaussian", penalty.factor = penaltyFactor) 
#> 
#>    Df  %Dev  Lambda
#> 1   0  0.00 0.54920
#> 2   1  5.69 0.50040
#> 3   1 10.41 0.45600
#> 4   1 14.34 0.41550
#> 5   1 17.59 0.37860
#> 6   1 20.30 0.34490
#> 7   1 22.54 0.31430
#> 8   1 24.40 0.28640
#> 9   1 25.95 0.26090
#> 10  1 27.23 0.23780
#> 11  1 28.30 0.21660
#> 12  1 29.19 0.19740
#> 13  1 29.92 0.17980
#> 14  1 30.53 0.16390
#> 15  2 31.12 0.14930
#> 16  2 32.40 0.13600
#> 17  2 33.46 0.12400
#> 18  2 34.33 0.11300
#> 19  3 35.08 0.10290
#> 20  4 35.83 0.09377
#> 21  4 36.98 0.08544
#> 22  4 37.94 0.07785
#> 23  4 38.73 0.07094
#> 24  4 39.39 0.06463
#> 25  4 39.94 0.05889
#> 26  4 40.40 0.05366
#> 27  4 40.78 0.04889
#> 28  4 41.09 0.04455
#> 29  4 41.35 0.04059
#> 30  4 41.57 0.03699
#> 31  4 41.75 0.03370
#> 32  4 41.90 0.03071
#> 33  4 42.02 0.02798
#> 34  4 42.12 0.02549
#> 35  4 42.21 0.02323
#> 36  4 42.28 0.02116
#> 37  4 42.34 0.01928
#> 38  4 42.39 0.01757
#> 39  4 42.43 0.01601
#> 40  4 42.46 0.01459
#> 41  4 42.49 0.01329
#> 42  4 42.51 0.01211
#> 43  4 42.53 0.01104
#> 44  4 42.55 0.01006
#> 45  5 42.56 0.00916
#> 46  5 42.57 0.00835
#> 47  5 42.58 0.00761
#> 48  5 42.59 0.00693
#> 49  5 42.60 0.00632
#> 50  5 42.60 0.00575
#> 51  5 42.61 0.00524
#> 52  5 42.61 0.00478
#> 53  5 42.61 0.00435
#> 54  5 42.62 0.00397
#> 55  5 42.62 0.00361
#> 56  5 42.62 0.00329
#> 57  5 42.62 0.00300
#> 58  5 42.62 0.00273
#> 59  5 42.63 0.00249
#> 60  5 42.63 0.00227
#> 61  5 42.63 0.00207
#> 62  5 42.63 0.00188
#> 63  5 42.63 0.00172
#> 64  5 42.63 0.00156
#> 65  5 42.63 0.00142