Setup network options

Setup network options, such as using weighted or unweighted degree, which centrality measure to use

networkOptions(
  method = "pearson",
  unweighted = TRUE,
  cutoff = 0,
  centrality = "degree",
  minDegree = 0,
  nCores = 1,
  transFun = function(x) x,
  min.degree = deprecated(),
  n.cores = deprecated(),
  trans.fun = deprecated()
)

Arguments

method

in case of correlation and covariance, which method to use.

unweighted

calculate degree using unweighted network.

cutoff

cuttoff value in network edges to trim the network.

centrality

centrality measure to use, currently only supports degree.

minDegree

minimum value that individual penalty weight can take.

nCores

number of cores to use, default to 1.

transFun

See details below.

min.degree

n.cores

trans.fun

The transFun argument takes a function definition that will apply a transformation to the penalty vector calculated from the degree. This transformation allows to change how the penalty is applied.

Value

a list of options

See also

Examples

networkOptions(unweighted = FALSE)
#> $method
#> [1] "pearson"
#> 
#> $unweighted
#> [1] FALSE
#> 
#> $cutoff
#> [1] 0
#> 
#> $centrality
#> [1] "degree"
#> 
#> $nCores
#> [1] 1
#> 
#> $minDegree
#> [1] 0
#> 
#> $transFun
#> function (x) 
#> x
#> <environment: 0x55da38fc2360>
#>