Download protein-protein interactions from STRING DB

Download protein-protein interactions from STRING DB

stringDBhomoSapiens(
  version = "11.0",
  scoreThreshold = 0,
  removeText = TRUE,
  score_threshold = deprecated(),
  remove.text = deprecated()
)

Arguments

version

version of the database to use

scoreThreshold

remove scores below threshold

removeText

remove text mining-based scores

score_threshold

remove.text

Value

a data.frame with rows representing an interaction between two proteins, and columns the count of scores above the given score_threshold

Examples

# \donttest{
stringDBhomoSapiens(scoreThreshold = 800)
#> Rows: 11759454 Columns: 16
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: " "
#> chr  (2): protein1, protein2
#> dbl (14): neighborhood, neighborhood_transferred, fusion, cooccurence, homol...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 96,566 × 3
#>    from                 to                   combined_score
#>    <chr>                <chr>                         <dbl>
#>  1 9606.ENSP00000000233 9606.ENSP00000341344            805
#>  2 9606.ENSP00000000233 9606.ENSP00000360532            980
#>  3 9606.ENSP00000000233 9606.ENSP00000440005            871
#>  4 9606.ENSP00000001008 9606.ENSP00000292169            858
#>  5 9606.ENSP00000001008 9606.ENSP00000354558            850
#>  6 9606.ENSP00000003084 9606.ENSP00000306330            857
#>  7 9606.ENSP00000003084 9606.ENSP00000358866            960
#>  8 9606.ENSP00000003084 9606.ENSP00000379287            857
#>  9 9606.ENSP00000003100 9606.ENSP00000370748            891
#> 10 9606.ENSP00000003100 9606.ENSP00000261507            980
#> # ℹ 96,556 more rows
# }